move BaseCommand to its own file
